+++ test/com/redhat/qe/auto/bugzilla/bz_checker_using_rest_tests.clj
@@ -23,7 +23,7 @@
     (is (= IBugzillaAPI$bzState/CLOSED (.getBugState checker "1")))
     (is (= "CLOSED" (.getBugField checker "1" "status")))
     (is (= "Bugzilla" (.getBugField checker "1" "product") ))
-    (is (= "bugzilla@redhat.com" (.getBugField checker "1" "qa_contact")))
+    (is (= "nobody@redhat.com" (.getBugField checker "1" "qa_contact")))
     (is (= #{"Reopened" "TestCaseApproved" "TestCaseRejected"}
            (into #{} (.getBugField checker "1" "keywords"))))
     (is (= false (.isBugOpen checker "1")))